requirements.txt
asteval==0.9.32
cachetools==5.3.3
click==8.1.7
dill==0.3.8
multiprocess==0.70.16
numpy==1.26.4
pnoise==0.2.0
pyphen==0.14.0
pyserial==3.5
scipy==1.13.0
shapely==2.0.4
svgelements==1.9.6
svgwrite==1.4.3
tomli==2.0.1
tqdm==4.66.4
vpype==1.14.0
vpype-gcode==0.13.0
